WIP on server manages physics
This is a large change, aiming at making it possible to properly
synchronise the game state (incl physics) between the peers.

The server is the only peer running any physics simulation at all:
players send their desired motion velocity, the server computes the
actual motion, and it broadcasts the resulting positions back to the
players. This applies similarly to the rigid bodies.

Current state is WIP:

- players can send their desired motion to the server

What is implemented but does not work:
- the server compute the motion
- the server boradcast the result position

-> cf Character.gd/_physics_process: for some reason, the code there is
not executed on the server (seems that is_network_master returns false,
even on the server)

What is not done yet:
- boardcast the position of the non-player rigid bodies
- transmit and broadcast the rotation of the player