WIP on proper character motion
The gravity is now appropriately handled by the server in _physics_process. This fix
issues where it was impossible to actually move a player after a few seconds of game play.

Still an issue with the gravity, though: once in the air, the characters do not fall back to
the floor.